refactor: remove unnecessary class and arrange package
