+++ src/main/java/nextstep/subway/line/domain/Line.java
@@ -14,7 +14,6 @@ import javax.persistence.Id;
 import javax.persistence.OneToMany;
 
 import nextstep.subway.common.BaseEntity;
-import nextstep.subway.section.domain.Section;
 import nextstep.subway.station.domain.Station;
 
 @Entity
@@ -35,6 +34,12 @@ public class Line extends BaseEntity {
 	public Line(String name, String color) {
 		this.name = name;
 		this.color = color;
+	}
+
+	public Line(String name, String color, Station upStation, Station downStation, int distance) {
+		this.name = name;
+		this.color = color;
+		sections.add(new Section(this, upStation, downStation, distance));
 	}
 
 	public void update(Line line) {
