+++ src/test/java/nextstep/subway/station/StationAcceptanceTest.java
@@ -1,6 +1,5 @@
 package nextstep.subway.station;
 
-import static nextstep.subway.station.StationAcceptanceTestHelper.*;
 import static org.assertj.core.api.Assertions.*;
 
 import java.util.HashMap;
@@ -103,5 +102,19 @@ public class StationAcceptanceTest extends AcceptanceTest {
 
 		// then
 		assertThat(response.statusCode()).isEqualTo(HttpStatus.NO_CONTENT.value());
+	}
+
+	public static ExtractableResponse<Response> 지하철역_생성_요청(String name) {
+		Map<String, String> params = new HashMap<>();
+		params.put("name", name);
+
+		// when
+		return RestAssured.given().log().all()
+			.body(params)
+			.contentType(MediaType.APPLICATION_JSON_VALUE)
+			.when()
+			.post("/stations")
+			.then().log().all()
+			.extract();
 	}
 }
